+++ gaios/WOFlow/Controllers/WOLoginViewController.swift
@@ -20,8 +20,9 @@ class WOLoginViewController: KeyboardViewController {
     var account: Account!
     private var buttonConstraint: NSLayoutConstraint?
     private var progressToken: NSObjectProtocol?
+    private let viewModel = WOViewModel()
     let menuButton = UIButton(type: .system)
-    var isSS = false
+    var isSS: Bool { account.gdkNetwork?.electrum ?? false }
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -142,46 +143,36 @@ class WOLoginViewController: KeyboardViewController {
         }
     }
 
+    func login() {
+        let password = self.passwordTextField.text ?? ""
+        firstly {
+            dismissKeyboard()
+            self.startLoader(message: NSLocalizedString("id_logging_in", comment: ""))
+            return Guarantee()
+        }.then { self.isSS ? self.viewModel.loginSinglesig(for: self.account) : self.viewModel.loginMultisig(for: self.account, password: password) }
+        .ensure { self.stopLoader() }
+        .done { _ = AccountNavigator.goLogged(account: self.account, nv: self.navigationController) }
+        .catch { error in
+            var prettyError = "id_login_failed"
+            switch error {
+            case TwoFactorCallError.failure(let localizedDescription):
+                prettyError = localizedDescription
+            case LoginError.connectionFailed:
+                prettyError = "id_connection_failed"
+            case LoginError.failed:
+                prettyError = "id_login_failed"
+            default:
+                break
+            }
+            DropAlert().error(message: NSLocalizedString(prettyError, comment: ""))
+            AnalyticsManager.shared.failedWalletLogin(account: self.account, error: error, prettyError: prettyError)
+            WalletsRepository.shared.delete(for: self.account)
+        }
+    }
+
     @objc func click(_ sender: Any) {
         view.endEditing(true)
-
-        if isSS {
-            print("login SS")
-        } else {
-            let username = self.account?.username ?? ""
-            let password = self.passwordTextField.text ?? ""
-            let bgq = DispatchQueue.global(qos: .background)
-            firstly {
-                view.endEditing(true)
-                self.startLoader(message: NSLocalizedString("id_logging_in", comment: ""))
-                return Guarantee()
-            }.compactMap {
-                WalletsRepository.shared.getOrAdd(for: self.account)
-            }.then(on: bgq) {
-                $0.login(Credentials(username: username, password: password))
-            }.ensure {
-                self.stopLoader()
-            }.done { _ in
-                let account = AccountsRepository.shared.current
-                AnalyticsManager.shared.loginWallet(loginType: .watchOnly, ephemeralBip39: false, account: account)
-                _ = AccountNavigator.goLogged(account: account!, nv: self.navigationController)
-            }.catch { error in
-                var prettyError = "id_login_failed"
-                switch error {
-                case TwoFactorCallError.failure(let localizedDescription):
-                    prettyError = localizedDescription
-                case LoginError.connectionFailed:
-                    prettyError = "id_connection_failed"
-                case LoginError.failed:
-                    prettyError = "id_login_failed"
-                default:
-                    break
-                }
-                DropAlert().error(message: NSLocalizedString(prettyError, comment: ""))
-                AnalyticsManager.shared.failedWalletLogin(account: self.account, error: error, prettyError: prettyError)
-                WalletsRepository.shared.delete(for: self.account)
-            }
-        }
+        login()
     }
 
     @IBAction func btnSettings(_ sender: Any) {