+++ CHANGELOG.md
@@ -3,7 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unrelease]
+## [Unreleased]
+
+### Added
+- Add singlesig Watch-Only login
 
 ### Fixed
 - Setup Pgp on multiple multisig networks